@@ -40,7 +40,6 @@ RUNNING_ANNOUNCEMENT = (
 
 
 class Admin:
-
     def __init__(self, config=Config):
         self.conf = config.get_conf(self, 8237492837454039, force_registration=True)
 
@@ -265,20 +264,16 @@ class Admin:
     @announce.command(name="ignore")
     @commands.guild_only()
     @checks.guildowner_or_permissions(administrator=True)
-    async def announce_ignore(self, ctx, *, guild: discord.Guild = None):
+    async def announce_ignore(self, ctx):
         """
-        Toggles whether the announcements will ignore the given server.
-            Defaults to the current server if none is provided.
+        Toggles whether the announcements will ignore the current server.
         """
-        if guild is None:
-            guild = ctx.guild
-
-        ignored = await self.conf.guild(guild).announce_ignore()
-        await self.conf.guild(guild).announce_ignore.set(not ignored)
+        ignored = await self.conf.guild(ctx.guild).announce_ignore()
+        await self.conf.guild(ctx.guild).announce_ignore.set(not ignored)
 
         verb = "will" if ignored else "will not"
 
-        await ctx.send("The server {} {} receive announcements.".format(guild.name, verb))
+        await ctx.send(f"The server {ctx.guild.name} {verb} receive announcements.")
 
     async def _valid_selfroles(self, guild: discord.Guild) -> Tuple[discord.Role]:
         """
@@ -303,6 +298,8 @@ class Admin:
         """
         Add a role to yourself that server admins have configured as
             user settable.
+
+        NOTE: The role is case sensitive!
         """
         # noinspection PyTypeChecker
         await self._addrole(ctx, ctx.author, selfrole)
@@ -311,6 +308,8 @@ class Admin:
     async def selfrole_remove(self, ctx: commands.Context, *, selfrole: SelfRole):
         """
         Removes a selfrole from yourself.
+
+        NOTE: The role is case sensitive!
         """
         # noinspection PyTypeChecker
         await self._removerole(ctx, ctx.author, selfrole)
@@ -320,6 +319,8 @@ class Admin:
     async def selfrole_add(self, ctx: commands.Context, *, role: discord.Role):
         """
         Add a role to the list of available selfroles.
+
+        NOTE: The role is case sensitive!
         """
         async with self.conf.guild(ctx.guild).selfroles() as curr_selfroles:
             if role.id not in curr_selfroles:
@@ -332,6 +333,8 @@ class Admin:
     async def selfrole_delete(self, ctx: commands.Context, *, role: SelfRole):
         """
         Removes a role from the list of available selfroles.
+
+        NOTE: The role is case sensitive!
         """
         async with self.conf.guild(ctx.guild).selfroles() as curr_selfroles:
             curr_selfroles.remove(role.id)
