@@ -25,9 +25,7 @@ if sys.platform == "linux":
 PYTHON_OK = sys.version_info >= (3, 5)
 INTERACTIVE_MODE = not len(sys.argv) > 1  # CLI flags = non-interactive
 
-INTRO = (
-    "==========================\n" "Red Discord Bot - Launcher\n" "==========================\n"
-)
+INTRO = "==========================\nRed Discord Bot - Launcher\n==========================\n"
 
 IS_WINDOWS = os.name == "nt"
 IS_MAC = sys.platform == "darwin"
@@ -383,7 +381,7 @@ def debug_info():
         + "User: {}\n".format(user_who_ran)
     )
     print(info)
-    exit(0)
+    sys.exit(0)
 
 
 def main_menu():
@@ -457,7 +455,7 @@ def main_menu():
 def main():
     if not PYTHON_OK:
         raise RuntimeError(
-            "Red requires Python 3.5 or greater. " "Please install the correct version!"
+            "Red requires Python 3.5 or greater. Please install the correct version!"
         )
     if args.debuginfo:  # Check first since the function triggers an exit
         debug_info()
